+++ catalogue_graph/src/ingestor/transformers/short_description.py
